APIv4 - Accept `match` param for Export action
This makes the Export action a little friendlier by inserting the `match`
param for you, if specified.
