+++ Civi/Api4/Generic/AbstractSaveAction.php
@@ -38,6 +38,7 @@ use Civi\Api4\Utils\CoreUtil;
  * @package Civi\Api4\Generic
  */
 abstract class AbstractSaveAction extends AbstractAction {
+  use Traits\MatchParamTrait;
 
   /**
    * Array of $ENTITIES to save.
@@ -70,20 +71,6 @@ abstract class AbstractSaveAction extends AbstractAction {
    * @var bool
    */
   protected $reload = FALSE;
-
-  /**
-   * Specify fields to match for update.
-   *
-   * Normally each record is either created or updated based on the presence of an `id`.
-   * Specifying `$match` fields will also perform an update if an existing $ENTITY matches all specified fields.
-   *
-   * Note: the fields named in this param should be without any options suffix (e.g. `my_field` not `my_field:name`).
-   * Any options suffixes in the $records will be resolved by the api prior to matching.
-   *
-   * @var array
-   * @optionsCallback getMatchFields
-   */
-  protected $match = [];
 
   /**
    * @throws \API_Exception
@@ -192,21 +179,6 @@ abstract class AbstractSaveAction extends AbstractAction {
   public function addDefault(string $fieldName, $defaultValue) {
     $this->defaults[$fieldName] = $defaultValue;
     return $this;
-  }
-
-  /**
-   * Options callback for $this->match
-   * @return array
-   */
-  protected function getMatchFields() {
-    return (array) civicrm_api4($this->getEntityName(), 'getFields', [
-      'checkPermissions' => FALSE,
-      'action' => 'get',
-      'where' => [
-        ['type', 'IN', ['Field', 'Custom']],
-        ['name', 'NOT IN', (array) CoreUtil::getInfoItem($this->getEntityName(), 'primary_key')],
-      ],
-    ], ['name']);
   }
 
 }
